working on occurrences migration.
